pyproject.toml
[tool.poetry]
name = "template-poetry"
version = "0.1.0"
description = "This project serves as a template for initialising new Python projects with Poetry, including basic configuration for dependency management and virtual environment setup. It is designed to be cloned for quick setup of new repositories on GitHub."
authors = ["ediakatos <ediakatos@mapaction.org>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^1.0.1"
rasterio = "^1.4.3"
rasterstats = "^0.20.0"
numpy = "^2.2.3"
affine = "^2.4.0"
pandas = "^2.2.3"
rioxarray = "^0.18.2"
xarray = "^2025.1.2"
boto3 = "^1.37.11"
python-dotenv = "^1.0.1"
smart-open = {extras = ["s3"], version = "^7.1.0"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pymarkdownlnt = "^0.9.18"
isort = "^5.13.2"
pytest = "^8.2.2"
tabulate = "^0.9.0"
nbqa = "^1.8.5"
black = "^24.4.2"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
